add test - testValueOf

+++ src/main/java/domain/interest_rate/InterestRate.java
@@ -23,7 +23,7 @@ public interface InterestRate {
 	 * 총 성장 계수 = (1 + 월 이자율) ^ 투자 기간(개월)
 	 * @param investPeriod 투자 기간
 	 */
-	double calTotalGrowthFactor(InvestPeriod investPeriod);
+	BigDecimal calTotalGrowthFactor(InvestPeriod investPeriod);
 
 	/**
 	 * 월 회차에 해당하는 성장 계수를 계산합니다.
